[mox] Remove topo dependency.
Add an `id` to CachedNode to handle its "attribute guards" needing an element-unique scope.

Allows us to inline mox-impl into mox.

Resolves #199.
