+++ dom/src/interfaces/html_element.rs
@@ -29,12 +29,13 @@ pub trait HtmlElementBuilder: ElementBuilder {
         contenteditable(bool)
     }
 
-    /// Forms a class of attributes, called custom data attributes, that allow
-    /// proprietary information to be exchanged between the HTML and its DOM
-    /// representation that may be used by scripts.
-    fn data(self, key: impl ToString, value: impl AsRef<str>) -> Self {
-        self.attribute(&format!("data-{}", key.to_string()), value.as_ref())
-    }
+    // TODO(#240) intern the data attribute strings so we can have 'static lifetime attribute names
+    // /// Forms a class of attributes, called custom data attributes, that allow
+    // /// proprietary information to be exchanged between the HTML and its DOM
+    // /// representation that may be used by scripts.
+    // fn data(self, key: impl ToString, value: impl AsRef<str>) -> Self {
+    //     self.attribute(&format!("data-{}", key.to_string()), value.as_ref())
+    // }
 
     attr_method! {
         /// The directionality of the element. It can have the following values:
