+++ mox/Cargo.toml
@@ -13,9 +13,14 @@ repository = "https://github.com/anp/moxie.git"
 authors = ["Adam Perry <lol@anp.lol>"]
 edition = "2018"
 
+[lib]
+proc-macro = true
+
 [dependencies]
-mox-impl = { path = "impl", version = "0.11.0"}
-topo = { path = "../topo", version = "0.13.1" }
+proc-macro2 = "1"
+quote = "1"
+syn-rsx = "0.8.0-beta.2"
+syn = "^1"
 
 [dev-dependencies]
 derive_builder = "0.9"
