--- mox/impl/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "mox-impl"
-version = "0.11.0"
-description = "Implementation for the mox! macro."
-categories = ["gui", "rust-patterns"]
-keywords = ["jsx", "xml", "builder"]
-readme = "../CHANGELOG.md"
-
-# update here, update everywhere!
-license = "MIT/Apache-2.0"
-homepage = "https://moxie.rs"
-repository = "https://github.com/anp/moxie.git"
-authors = ["Adam Perry <lol@anp.lol>"]
-edition = "2018"
-
-[lib]
-proc-macro = true
-
-[dependencies]
-proc-macro2 = "1"
-quote = "1"
-syn-rsx = "0.8.0-beta.2"
-syn = "^1"
